feat(symfony): require SF6.4+ & PHP8.1+
