+++ composer.json
@@ -12,7 +12,7 @@
         }
     ],
     "require": {
-        "php": ">=8.0",
+        "php": ">=8.1",
         "league/flysystem": "^3.11",
         "zenstruck/image": "^0.5.0",
         "zenstruck/stream": "^1.2",
@@ -32,17 +32,17 @@
         "league/mime-type-detection": ">=1.8",
         "phpstan/phpstan": "^1.4",
         "phpunit/phpunit": "^9.5.0",
-        "symfony/browser-kit": "^5.4|^6.0|^7.0",
-        "symfony/form": "^5.4|^6.0|^7.0",
-        "symfony/framework-bundle": "^5.4|^6.0|^7.0",
-        "symfony/mime": "^5.4|^6.0|^7.0",
+        "symfony/browser-kit": "^6.4|^7.0",
+        "symfony/form": "^6.4|^7.0",
+        "symfony/framework-bundle": "^6.4|^7.0",
+        "symfony/mime": "^6.4|^7.0",
         "symfony/phpunit-bridge": "^6.1|^7.0",
-        "symfony/serializer": "^5.4|^6.0|^7.0",
-        "symfony/stopwatch": "^5.4|^6.0|^7.0",
-        "symfony/string": "^5.4|^6.0|^7.0",
-        "symfony/twig-bundle": "^5.4|^6.0|^7.0",
-        "symfony/validator": "^5.4|^6.0|^7.0",
-        "symfony/var-dumper": "^5.4|^6.0|^7.0",
+        "symfony/serializer": "^6.4|^7.0",
+        "symfony/stopwatch": "^6.4|^7.0",
+        "symfony/string": "^6.4|^7.0",
+        "symfony/twig-bundle": "^6.4|^7.0",
+        "symfony/validator": "^6.4|^7.0",
+        "symfony/var-dumper": "^6.4|^7.0",
         "zenstruck/assert": "^1.2",
         "zenstruck/console-test": "^1.4",
         "zenstruck/foundry": "^1.30.3",
