+++ src/Filesystem/Glide/GlideTransformUrlGenerator.php
@@ -28,7 +28,7 @@ final class GlideTransformUrlGenerator implements TransformUrlGenerator
     {
         $filter = match (true) {
             \is_string($filter) => ['p' => $filter], // is glide "preset"
-            \is_array($filter) && !array_is_list($filter) => $filter, // is standard glide parameters
+            \is_array($filter) && !\array_is_list($filter) => $filter, // is standard glide parameters
             \is_array($filter) => ['p' => \implode(',', $filter)], // is array of "presets"
         };
